+++ src/components/function.go
@@ -7,14 +7,16 @@ import (
 	"log"
 	"math"
 	"os"
-	"os/user"
 	"path/filepath"
 	"regexp"
 	"sort"
 	"strconv"
 	"strings"
+	"time"
 
+	"github.com/lithammer/shortuuid"
 	"github.com/rkoesters/xdg/userdirs"
+	"github.com/shirou/gopsutil/disk"
 )
 
 func getDirectories() []directory {
@@ -89,7 +91,30 @@ func getExternalMediaFolders() []directory {
 	return directories
 }
 
-// TODO: Remove this function
+func GetExternalDisk() (disks []disk.PartitionStat, err error) {
+	parts, err := disk.Partitions(true)
+
+	if err != nil {
+		return []disk.PartitionStat{}, err
+	}
+	for _, disk := range parts {
+		if IsExternalDiskPath(disk.Mountpoint) {
+		disks = append(disks, disk)
+		}
+	}
+
+	return disks, err
+}
+
+func IsExternalDiskPath(path string) bool {
+	dir := filepath.Dir(path)
+	return strings.HasPrefix(dir, "/mnt") ||
+		strings.HasPrefix(dir, "/media") ||
+		strings.HasPrefix(dir, "/run/media") ||
+		strings.HasPrefix(dir, "/Volumes")
+}
+
+  // TODO: Remove this function
 // This gets the award for most redundant function seen by @lescx ever
 func repeatString(s string, count int) string {
 	return strings.Repeat(s, count)
@@ -320,7 +345,7 @@ func PasteDir(src, dst string, id string, m model) (model, error) {
 
 			if len(channel) < 5 {
 				channel <- channelMessage{
-					processId:       id,
+					messageId:       id,
 					processNewState: p,
 				}
 			}
@@ -329,7 +354,7 @@ func PasteDir(src, dst string, id string, m model) (model, error) {
 			if err != nil {
 				p.state = failure
 				channel <- channelMessage{
-					processId:       id,
+					messageId:       id,
 					processNewState: p,
 				}
 				return err
@@ -337,7 +362,7 @@ func PasteDir(src, dst string, id string, m model) (model, error) {
 			p.done++
 			if len(channel) < 5 {
 				channel <- channelMessage{
-					processId:       id,
+					messageId:       id,
 					processNewState: p,
 				}
 			}
@@ -365,17 +390,31 @@ func contains(s []string, str string) bool {
 
 func ReturnMetaData(m model) model {
 	panel := m.fileModel.filePanels[m.filePanelFocusIndex]
+	cursor := panel.cursor
+	LastTimeCursorMove = [2]int{int(time.Now().UnixMicro()), cursor}
+	time.Sleep(150 * time.Millisecond)
+	if LastTimeCursorMove[1] != cursor && m.focusPanel != metaDataFocus {
+		return m
+	}
+	m.fileMetaData.metaData = m.fileMetaData.metaData[:0]
+	id := shortuuid.New()
 	if len(panel.element) == 0 {
+		channel <- channelMessage{
+			messageId:    id,
+			loadMetadata: true,
+			metadata:     m.fileMetaData.metaData,
+		}
 		return m
 	}
 	if len(panel.element[panel.cursor].metaData) != 0 && m.focusPanel != metaDataFocus {
 		m.fileMetaData.metaData = panel.element[panel.cursor].metaData
+		channel <- channelMessage{
+			messageId:    id,
+			loadMetadata: true,
+			metadata:     m.fileMetaData.metaData,
+		}
 		return m
 	}
-	if len(panel.element) == 0 {
-		return m
-	}
-	m.fileMetaData.metaData = m.fileMetaData.metaData[:0]
 	filePath := panel.element[panel.cursor].location
 
 	fileInfo, err := os.Stat(filePath)
@@ -392,6 +431,11 @@ func ReturnMetaData(m model) model {
 			m.fileMetaData.metaData = append(m.fileMetaData.metaData, [2]string{"FolderSize", FormatFileSize(DirSize(filePath))})
 		}
 		m.fileMetaData.metaData = append(m.fileMetaData.metaData, [2]string{"FolderModifyDate", fileInfo.ModTime().String()})
+		channel <- channelMessage{
+			messageId:    id,
+			loadMetadata: true,
+			metadata:     m.fileMetaData.metaData,
+		}
 		return m
 	}
 
@@ -407,6 +451,12 @@ func ReturnMetaData(m model) model {
 			m.fileMetaData.metaData = append(m.fileMetaData.metaData, temp)
 		}
 	}
+	channel <- channelMessage{
+		messageId:    id,
+		loadMetadata: true,
+		metadata:     m.fileMetaData.metaData,
+	}
+
 	panel.element[panel.cursor].metaData = m.fileMetaData.metaData
 	return m
 }
@@ -452,8 +502,4 @@ func countFiles(dirPath string) (int, error) {
 	})
 
 	return count, err
-}
-
-func IsExternalPath(path string) bool {
-	return strings.HasPrefix(path, "/run/media")
 }